Improve GUI React Doctor score
