+++ .gitignore
@@ -9,3 +9,4 @@ tests/evals/runs/*
 graphify-out/
 tests/screenshots/out/
 website/dist/
+validation-output/
