+++ gui/web/src/components/ui/source-pill.jsx
@@ -1,6 +1,7 @@
 import * as Tooltip from "@radix-ui/react-tooltip";
 import { ExternalLink } from "lucide-react";
-import { Favicon, hostFrom, shortHost } from "./favicon.jsx";
+import { Favicon } from "./favicon.jsx";
+import { hostFrom, shortHost } from "./favicon-utils.js";
 import { cn } from "../../lib/utils.js";
 
 // Compact pill: favicon + hostname. Clickable when given a url; the surrounding
@@ -48,33 +49,5 @@ export function SourcePill({ url, label, title, snippet, className, onClick }) {
         </Tooltip.Portal>
       </Tooltip.Root>
     </Tooltip.Provider>
-  );
-}
-
-export function SourcePillRow({ sources, max = 12, className, onOverflowClick }) {
-  if (!sources?.length) return null;
-  const shown = sources.slice(0, max);
-  const overflow = sources.length - shown.length;
-  return (
-    <div className={cn("flex flex-wrap gap-1.5", className)}>
-      {shown.map((s, i) => (
-        <SourcePill
-          key={`${s.url || s.title || i}`}
-          url={s.url}
-          label={s.label || s.source}
-          title={s.title}
-          snippet={s.snippet}
-        />
-      ))}
-      {overflow > 0 ? (
-        <button
-          type="button"
-          onClick={onOverflowClick}
-          className="inline-flex items-center gap-1.5 rounded-full border border-border bg-secondary px-2 py-0.5 text-[11px] text-muted-foreground transition-colors hover:bg-card hover:text-foreground focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-ring focus-visible:ring-offset-2 focus-visible:ring-offset-background"
-        >
-          +{overflow} more
-        </button>
-      ) : null}
-    </div>
   );
 }
