+++ gui/web/src/features/catalog/form-presets.js
@@ -1,0 +1,37 @@
+export const RANGE_OPTIONS = [
+  { value: "1mo", label: "1M" },
+  { value: "3mo", label: "3M" },
+  { value: "6mo", label: "6M" },
+  { value: "1y", label: "1Y" },
+  { value: "2y", label: "2Y" },
+  { value: "5y", label: "5Y" },
+  { value: "max", label: "Max" },
+];
+
+export const PERIOD_OPTIONS = [
+  { value: "6mo", label: "6M" },
+  { value: "1y", label: "1Y" },
+  { value: "2y", label: "2Y" },
+];
+
+export const FRESHNESS_OPTIONS = [
+  { value: "hours", label: "Hours" },
+  { value: "day", label: "Day" },
+  { value: "week", label: "Week" },
+  { value: "month", label: "Month" },
+];
+
+export const HOURS_PRESETS = [
+  { value: 6, label: "6h" },
+  { value: 24, label: "24h" },
+  { value: 72, label: "3d" },
+  { value: 168, label: "7d" },
+];
+
+export const DAYS_PRESETS = [
+  { value: 7, label: "7d" },
+  { value: 30, label: "30d" },
+  { value: 90, label: "90d" },
+  { value: 180, label: "6M" },
+  { value: 365, label: "1Y" },
+];
