+++ gui/web/src/features/renderers/ToolDrawerStep.jsx
@@ -2,7 +2,8 @@ import { CheckCircle2, AlertCircle } from "lucide-react";
 import { StatusDot } from "../../components/ui/status-dot.jsx";
 import { TextShimmer } from "../../components/ui/text-shimmer.jsx";
 import { GenericCard, rendererFor } from "./cards/index.jsx";
-import { PlainOutput, RawDetails, extractDetails } from "./cards/_shared.jsx";
+import { PlainOutput, RawDetails } from "./cards/_shared.jsx";
+import { extractDetails } from "./cards/card-format.js";
 import { ToolIcon, toolMeta } from "./tool-icon.jsx";
 import { textContent } from "../../rendering/text.js";
 import { cn } from "../../lib/utils.js";
