+++ gui/web/src/features/renderers/cards/macro.jsx
@@ -1,5 +1,6 @@
 import { Badge } from "../../../components/ui/badge.jsx";
-import { DeltaChip, MoneyTile, PlainOutput, RangeBar, Sparkline, StatRow, ToolCard, extractDetails, formatDateShort } from "./_shared.jsx";
+import { DeltaChip, MoneyTile, PlainOutput, RangeBar, Sparkline, StatRow, ToolCard } from "./_shared.jsx";
+import { extractDetails, formatDateShort } from "./card-format.js";
 
 export function MacroSeriesCard({ message, header, text }) {
   const d = extractDetails(message);
